requirements.txt
Sphinx==1.1.3
nose==1.3.0
PyYAML==3.10
